requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
openpyxl==3.1.2
reportlab==4.0.4
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
PyMySQL==1.1.0
mysqlclient==2.2.0